+++ packages/winnow/test/to-esri.js
@@ -1,6 +1,7 @@
 const test = require('tape')
 const geojson = require('./fixtures/to-esri-fixture.json')
 const Winnow = require('../src')
+const _ = require('lodash')
 
 test('detecting fields', t => {
   const options = {}
@@ -13,11 +14,36 @@ test('detecting fields', t => {
   t.end()
 })
 
+test('adding an object id', t => {
+  const options = {
+    toEsri: true
+  }
+  const fixture = _.cloneDeep(geojson)
+  const result = Winnow.query(fixture, options)
+  t.equal(result.features[0].attributes.OBJECTID, 0)
+  t.end()
+})
+
+test('do not overwrite the object id', t => {
+  const options = {
+    toEsri: true
+  }
+
+  const fixture = _.cloneDeep(geojson)
+  fixture.features[0].properties.OBJECTID = 1
+  fixture.metadata = { idField: 'OBJECTID' }
+
+  const result = Winnow.query(fixture, options)
+  t.equal(result.features[0].attributes.OBJECTID, 1)
+  t.end()
+})
+
 test('converting date fields', t => {
   const options = {
     toEsri: true
   }
-  const result = Winnow.query(geojson, options)
+  const fixture = _.cloneDeep(geojson)
+  const result = Winnow.query(fixture, options)
   t.equal(result.features[0].attributes.date, 1331769600000)
   t.equal(Object.keys(result.metadata.fields).length, 4)
   t.end()